+++ src/main/java/de/samply/db/repository/ProjectRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -18,9 +19,14 @@ public interface ProjectRepository extends JpaRepository<Project, Long> {
     Optional<Project> findByCode(String projectCode);
 
 
+    @Query("SELECT p FROM Project p WHERE p.expiresAt < :expirationTime AND p.state IN :states")
+    List<Project> findByExpiresAtBeforeAndStateIn(LocalDate expirationTime, Set<ProjectState> states);
+
     ////////// Project Manager Admins:
     List<Project> findAll();
+
     Page<Project> findAllByOrderByModifiedAtDesc(Pageable pageable);
+
     Page<Project> findAllByOrderByModifiedAtAsc(Pageable pageable);
 
     @Query("SELECT p FROM Project p WHERE p.archivedAt IS NOT NULL ORDER BY p.modifiedAt DESC")